requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.30.0
openai==1.35.0
google-generativeai==0.7.2
python-dotenv==1.0.1
chromadb==0.5.3
pypdf
python-pptx==0.6.23
reportlab==4.2.2
requests==2.32.3
tiktoken==0.7.0
httpx==0.27.2
